refactor: ProductDetailCard 컴포넌트 제거
ProductDetailCard의 내용을 ProductDetailPage로 이동
