+++ src/pages/ProductDetailPage/index.jsx
@@ -5,7 +5,16 @@ import { useProductDetail } from "../../hooks/useProductDetail";
 import { useCartItemList } from "../../hooks/useCartItemList";
 
 import Spinner from "../../components/common/Spinner";
-import ProductDetailCard from "./ProductDetailCard";
+import Button from "../../components/common/Button";
+import {
+  StyledBottomSection,
+  StyledContainer,
+  StyledProductImg,
+  StyledProductName,
+  StyledProductPrice,
+  StyledProductPriceText,
+  StyledTopSection,
+} from "./index.styled";
 
 function ProductDetailPage() {
   const { id: productId } = useParams();
@@ -38,15 +47,41 @@ function ProductDetailPage() {
     cartItemListIndex === -1 ? 0 : cartItemList[cartItemListIndex].quantity;
 
   return (
-    <ProductDetailCard
-      product={{ ...product, quantity }}
-      onClickAddToCartButton={() => {
-        updateCartItemQuantityWithAlert({
-          id: product.id,
-          quantity: quantity + 1,
-        });
-      }}
-    />
+    <StyledContainer>
+      <StyledTopSection>
+        <StyledProductImg src={product.thumbnailUrl} alt={product.name} />
+        <StyledProductName>{product.name}</StyledProductName>
+      </StyledTopSection>
+      <StyledBottomSection>
+        <StyledProductPriceText>금액</StyledProductPriceText>
+        <StyledProductPrice>
+          {product.price.toLocaleString()}원
+        </StyledProductPrice>
+      </StyledBottomSection>
+      <AddToCartButton
+        onClickAddToCartButton={() => {
+          updateCartItemQuantityWithAlert({
+            id: product.id,
+            quantity: quantity + 1,
+          });
+        }}
+      />
+    </StyledContainer>
+  );
+}
+
+function AddToCartButton({ ...props }) {
+  return (
+    <Button
+      height="60px"
+      fontSize="1.25rem"
+      fontWeight="700"
+      color="white"
+      bgColor="brown"
+      {...props}
+    >
+      장바구니 담기
+    </Button>
   );
 }
